--- swanlab/api/__init__.py
@@ -1,20 +1,0 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-r"""
-@DATE: 2024/4/3 21:52
-@File: __init__.py.py
-@IDE: pycharm
-@Description:
-    API模块，封装api请求接口
-"""
-from .auth.login import terminal_login, code_login
-from .info import *
-
-
-__all__ = [
-    "LoginInfo",
-    'ExperimentInfo',
-    'ProjectInfo',
-    "code_login",
-    'terminal_login',
-]
