+++ swanlab/core_python/auth/__init__.py
@@ -1,0 +1,10 @@
+"""
+@author: cunyue
+@file: __init__.py
+@time: 2025/6/19 15:09
+@description:
+"""
+
+from .providers.api_key import *
+
+# NOTE: 未来的其他认证方式应该使用模块的方式导入，比如 from .providers import oauth, api_key 是最基础的认证方式，因此将相关函数全部导出
