+++ swanlab/core_python/auth/providers/api_key.py
@@ -1,25 +1,102 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-r"""
-@DATE: 2024-03-19 15:30:24
-@File: swanlab/auth/login.py
-@IDE: vscode
-@Description:
-    用户登录接口，输入用户的apikey，保存用户token到本地
-    进行一些交互定义和数据请求
 """
+@author: cunyue
+@file: session.py
+@time: 2025/6/19 14:52
+@description: 使用api_key登录后获取会话凭证
+"""
+
 import getpass
 import sys
+from typing import Union
 
 import requests
 from swankit.env import is_windows
 from swankit.log import FONT
 
-from swanlab.api.info import LoginInfo
 from swanlab.env import in_jupyter
 from swanlab.error import ValidationError, APIKeyFormatError
 from swanlab.log import swanlog
-from swanlab.package import get_setting_url, get_host_api, get_host_web
+from swanlab.package import get_setting_url, get_host_api, get_host_web, fmt_web_host, save_key as sk
+
+
+class LoginInfo:
+    """
+    登录信息类，负责解析登录接口返回的信息，并且进行保存
+    无论接口请求成功还是失败，都会初始化一个LoginInfo对象
+    """
+
+    def __init__(self, resp: requests.Response, api_key: str, api_host: str, web_host: str):
+        self.__api_key = api_key
+        self.__resp = resp
+        self.__body = resp.json() if resp.status_code == 200 else {}
+        self.__username = None
+        """
+        如果此属性不为None，username返回此属性
+        """
+        self.web_host = fmt_web_host(web_host)
+        self.api_host = api_host
+
+    @property
+    def sid(self) -> Union[str, None]:
+        """
+        获取sid，如果请求失败则返回None
+        """
+        return self.__body.get("sid")
+
+    @property
+    def expired_at(self) -> Union[str, None]:
+        """
+        获取过期时间，如果请求失败则返回None
+        """
+        return self.__body.get("expiredAt")
+
+    @property
+    def username(self) -> Union[str, None]:
+        """
+        获取用户名，如果请求失败则返回None
+        """
+        if self.__username is not None:
+            return self.__username
+        return self.__body.get("userInfo", {}).get("username")
+
+    @username.setter
+    def username(self, value):
+        """
+        设置用户名
+        """
+        self.__username = value
+
+    @property
+    def is_fail(self):
+        """
+        判断登录是否失败
+        """
+        return self.__resp.status_code != 200
+
+    @property
+    def api_key(self):
+        """
+        获取api_key
+        """
+        if self.is_fail:
+            return None
+        return self.__api_key
+
+    def __str__(self) -> str:
+        """错误时会返回错误信息"""
+        if self.__resp.reason == "OK":
+            return "Login success"
+        if self.__resp.reason == "Unauthorized" or self.__resp.reason == "Authorization Required":
+            return "Error api key"
+        if self.__resp.reason == "Forbidden":
+            return "You need to be verified first"
+        return str(self.__resp.status_code) + " " + self.__resp.reason
+
+    def save(self):
+        """
+        保存登录信息
+        """
+        return sk(self.web_host, self.api_key, self.api_host)
 
 
 def login_request(api_key: str, api_host: str, timeout: int = 20) -> requests.Response:
@@ -132,3 +209,12 @@ def _abort_tip(tp, _, __):
     if tp == KeyboardInterrupt:
         sys.exit(0)
     # 如果不是CTRL+C，交给默认的异常处理
+
+
+__all__ = [
+    "login_request",
+    "login_by_key",
+    "terminal_login",
+    "code_login",
+    "LoginInfo",
+]
