+++ test/unit/core_python/test_client.py
@@ -14,8 +14,8 @@ import responses
 from responses import registries
 from swankit.core import MediaBuffer
 
-from swanlab.api.auth.login import login_by_key
 from swanlab.core_python import create_client, Client, CosClient
+from swanlab.core_python.auth import login_by_key
 from swanlab.package import get_host_api
 from tutils import is_skip_cloud_test, TEMP_PATH, API_KEY
 from tutils.setup import *
